refactor(design): Phase 1 ポートIF整理 — 逆依存解消・契約の正規化
- BufferedEventStore IF を application → shared/ports.ts に移動し、
  infrastructure → application の逆依存を解消（関心の分離）
- SessionStorePort を shared/types.ts に追加し、RunnerDeps と各
  エージェント Deps をポート型に変更（契約による設計）
- MemoryReadServices IF を mcp/tools → memory/index.ts に昇格し、
  読み取り専用サブセットを公開型として再利用可能に（カプセル化）
