+++ packages/infrastructure/src/store/sqlite-buffered-event-store.ts
@@ -1,4 +1,4 @@
-import type { BufferedEventStore } from "@vicissitude/application/message-ingestion-service";
+import type { BufferedEventStore } from "@vicissitude/shared/ports";
 import type { BufferedEvent } from "@vicissitude/shared/types";
 import type { StoreDb } from "@vicissitude/store/db";
 import { appendEvent } from "@vicissitude/store/queries";
